refactor ControlProportionSettings into reusable helper
