+++ public/src/components/channelManagement/helpers/controlProportionSettings.ts
@@ -1,0 +1,4 @@
+export interface ControlProportionSettings {
+  proportion: number;
+  offset: number;
+}
